refactor : message 폴더 추가
