+++ src/hooks/useCardNumbers.ts
@@ -8,7 +8,7 @@ import {
   CardNumbersOptions,
   IsError,
 } from '../types/CardNumbers';
-import { COMMON_ERROR_MESSAGE } from './commonErrorMessage';
+import { COMMON_ERROR_MESSAGE } from './message/commonErrorMessage';
 
 const INITIAL_CARD_NUMBER: CardNumbers = {
   firstNumber: '',
